Tone down blue from #6c9ef8 to #6693e8
6c9ef8 used to be converted to ctermfg=111 by colortemplate, which
works, but is uuuugly. On my own box I always knocked it down to
ctermfg=68, which felt like a much better approximation.

I also got a new screen recently, which I believe to be considerably
more accurate than the one I had before, and 6c9ef8 stands out way too
much.

I understand that in 256c the `Statement` and `Preproc` colors might now
be too close, but that's only really been an issue for me when editing
shell scripts, for which I usually load a barebones comments-only
colorscheme anyway. I'm willing to entertain suggestions as for what the
`Preproc` color could be changed to.